Add a function that will automagically load data from an HTTP URL for you
To see this in action, start the Marz development server, open the
main page, enter the JavaScript console and run the helper function:
waitForRenderAndLoad()

The core load function, loadHttp, is attached to the SidebarController,
so the scope of the #sidebarDrop element on the web page is required to
invoke it.

Note that this function is not robust against CORS errors; the data
file needs to come from the same domain hosting the Marz deployment.